refactor: Change style.ts to Component.style.ts
