+++ src/components/PostCard/PostCard.style.ts
